Moved MainMenuActivity and its viewmodel to one package
