+++ collect_app/src/main/AndroidManifest.xml
@@ -101,7 +101,7 @@ the specific language governing permissions and limitations under the License.
         </activity>
 
         <activity
-            android:name=".activities.MainMenuActivity"
+            android:name=".mainmenu.MainMenuActivity"
             android:theme="@style/Theme.Collect.SplashScreen"
             android:exported="true">
             <intent-filter>
