+++ collect_app/src/main/java/org/odk/collect/android/activities/CrashHandlerActivity.kt
@@ -2,6 +2,7 @@ package org.odk.collect.android.activities
 
 import android.os.Bundle
 import androidx.activity.OnBackPressedCallback
+import org.odk.collect.android.mainmenu.MainMenuActivity
 import org.odk.collect.crashhandler.CrashHandler
 import org.odk.collect.strings.localization.LocalizedActivity
