+++ collect_app/src/main/java/org/odk/collect/android/preferences/screens/UserInterfacePreferencesFragment.java
@@ -24,8 +24,8 @@ import android.content.Context;
 import android.os.Bundle;
 import androidx.preference.ListPreference;
 import org.odk.collect.android.R;
-import org.odk.collect.android.activities.MainMenuActivity;
 import org.odk.collect.android.injection.DaggerUtils;
+import org.odk.collect.android.mainmenu.MainMenuActivity;
 import org.odk.collect.android.utilities.LocaleHelper;
 import org.odk.collect.android.version.VersionInformation;
 import java.util.ArrayList;
