+++ collect_app/src/test/java/org/odk/collect/android/projects/ManualProjectCreatorDialogTest.kt
@@ -22,8 +22,8 @@ import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.verify
 import org.odk.collect.android.R
-import org.odk.collect.android.activities.MainMenuActivity
 import org.odk.collect.android.injection.config.AppDependencyModule
+import org.odk.collect.android.mainmenu.MainMenuActivity
 import org.odk.collect.android.support.CollectHelpers
 import org.odk.collect.android.support.Matchers.isPasswordHidden
 import org.odk.collect.fragmentstest.FragmentScenarioLauncherRule
